components/runners/ambient-runner/ambient_runner/endpoints/run.py
"""POST / — AG-UI run endpoint (delegates to bridge)."""

import logging
import uuid
from typing import Any, Dict, List, Optional, Union

from ag_ui.core import EventType, RunAgentInput, RunErrorEvent, ToolCallResultEvent
from ag_ui_claude_sdk.utils import now_ms
from ag_ui.encoder import EventEncoder
from fastapi import APIRouter, Request
from fastapi.responses import StreamingResponse
from pydantic import BaseModel

logger = logging.getLogger(__name__)

router = APIRouter()


class RunnerInput(BaseModel):
    """Input model with optional AG-UI fields."""

    threadId: Optional[str] = None
    thread_id: Optional[str] = None
    runId: Optional[str] = None
    run_id: Optional[str] = None
    parentRunId: Optional[str] = None
    parent_run_id: Optional[str] = None
    messages: List[Dict[str, Any]]
    state: Optional[Dict[str, Any]] = None
    tools: Optional[List[Any]] = None
    context: Optional[Union[List[Any], Dict[str, Any]]] = None
    forwardedProps: Optional[Dict[str, Any]] = None
    environment: Optional[Dict[str, str]] = None
    metadata: Optional[Dict[str, Any]] = None

    def to_run_agent_input(self) -> RunAgentInput:
        thread_id = self.threadId or self.thread_id
        run_id = self.runId or self.run_id or str(uuid.uuid4())
        parent_run_id = self.parentRunId or self.parent_run_id
        context_list = self.context if isinstance(self.context, list) else []

        return RunAgentInput(
            thread_id=thread_id,
            run_id=run_id,
            parent_run_id=parent_run_id,
            messages=self.messages,
            state=self.state or {},
            tools=self.tools or [],
            context=context_list,
            forwarded_props=self.forwardedProps or {},
        )


@router.post("/")
async def run_agent(input_data: RunnerInput, request: Request):
    """AG-UI run endpoint — delegates to the bridge."""
    bridge = request.app.state.bridge

    run_agent_input = input_data.to_run_agent_input()
    accept_header = request.headers.get("accept", "text/event-stream")
    encoder = EventEncoder(accept=accept_header)

    logger.info(
        f"Run: thread_id={run_agent_input.thread_id}, run_id={run_agent_input.run_id}"
    )

    async def event_stream():
        try:
            async for event in bridge.run(run_agent_input):
                try:
                    yield encoder.encode(event)
                except Exception as encode_err:
                    # A single event failed to encode (e.g. tool result > 1MB).
                    # Emit a fallback for that event and keep the run alive.
                    logger.warning(
                        "Failed to encode %s event: %s",
                        type(event).__name__,
                        encode_err,
                    )
                    tool_call_id = getattr(event, "tool_call_id", None)
                    if tool_call_id:
                        # Replace the oversized result with an error result
                        # so the tool call closes out in the UI.
                        fallback = ToolCallResultEvent(
                            type=EventType.TOOL_CALL_RESULT,
                            thread_id=getattr(event, "thread_id", "") or "",
                            run_id=getattr(event, "run_id", "") or "",
                            message_id=f"{tool_call_id}-result",
                            tool_call_id=tool_call_id,
                            role="tool",
                            content=(
                                f"[Tool result too large to display: {encode_err}]"
                            ),
                        )
                        yield encoder.encode(fallback)
                    else:
                        # Non-tool event too large (e.g. MessagesSnapshot).
                        # Emit a RunError so the frontend knows something
                        # was dropped rather than silently losing data.
                        yield encoder.encode(
                            RunErrorEvent(
                                type=EventType.RUN_ERROR,
                                thread_id=getattr(event, "thread_id", "") or run_agent_input.thread_id or "",
                                run_id=getattr(event, "run_id", "") or run_agent_input.run_id or "unknown",
                                message=f"An event was too large to send ({type(event).__name__}: {encode_err})",
                                timestamp=now_ms(),
                            )
                        )
        except Exception as e:
            logger.error(f"Error in event stream: {e}", exc_info=True)

            error_msg = str(e)
            extra = bridge.get_error_context()
            if extra:
                error_msg = f"{error_msg}\n\n{extra}"

            yield encoder.encode(
                RunErrorEvent(
                    type=EventType.RUN_ERROR,
                    thread_id=run_agent_input.thread_id or "",
                    run_id=run_agent_input.run_id or "unknown",
                    message=error_msg,
                    timestamp=now_ms(),
                )
            )

    return StreamingResponse(
        event_stream(),
        media_type=encoder.get_content_type(),
        headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
    )
