components/runners/ambient-runner/pyproject.toml
[project]
name = "ambient-runner"
version = "0.4.0"
description = "Polymorphic AG-UI server — Claude Code, Gemini CLI, Codex SDK"
readme = "CLAUDE.md"
requires-python = ">=3.11"
authors = [
  { name = "vTeam" }
]
dependencies = [
  # Ambient Runner SDK core
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.23.0",
  "ag-ui-protocol>=0.1.13",
  "pydantic>=2.0.0",
  "aiohttp>=3.8.0",
  "requests>=2.31.0",
  "pyjwt>=2.8.0",
]

[project.optional-dependencies]
claude = [
  "anthropic[vertex]>=0.68.0",
  "claude-agent-sdk>=0.1.23",
]
observability = [
  "langfuse>=3.0.0",
]
mcp-atlassian = [
  "mcp-atlassian>=0.11.9",
]
# Install everything for the polymorphic runner
all = [
  "ambient-runner[claude,observability,mcp-atlassian]",
]

[tool.uv]
dev-dependencies = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.1.0",
  "black>=23.0.0",
  "httpx>=0.24.0",
]

[tool.setuptools]
py-modules = ["main"]
packages = ["ag_ui_claude_sdk", "ag_ui_gemini_cli", "ambient_runner", "ambient_runner.endpoints", "ambient_runner.middleware", "ambient_runner.platform", "ambient_runner.bridges", "ambient_runner.bridges.claude", "ambient_runner.bridges.langgraph", "ambient_runner.bridges.gemini_cli"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
